PR:		197915
Reminded by:	dchagin

+++ sys/compat/cheriabi/cheriabi_sysstubs.h
@@ -3069,14 +3069,14 @@ SYS_STUB_VA(499, int, openat, flag,
     /* _localcheck */ {if (!(cheri_getperm(path) & CHERI_PERM_GLOBAL)) {errno = EPROT; return ((int)-1);} }
 )
 
-SYS_STUB(500, int, readlinkat,
+SYS_STUB(500, ssize_t, readlinkat,
     /* _protoargs */ (int fd, const char *  path, char *  buf, size_t bufsize),
-    /* _protoargs_chk */ (int *retp , int * __capability stub_errno, int fd, const char * __capability   path, char * __capability   buf, size_t bufsize),
+    /* _protoargs_chk */ (ssize_t *retp , int * __capability stub_errno, int fd, const char * __capability   path, char * __capability   buf, size_t bufsize),
     /* _protoargs_err */ (int * __capability stub_errno, int fd, const char * __capability   path, char * __capability   buf, size_t bufsize),
     /* _callargs */ (fd, (__cheri_fromcap const char * )path, (__cheri_fromcap char * )buf, bufsize),
     /* _callargs_chk */ (&ret, stub_errno, fd, path, buf, bufsize),
     /* _callargs_err */ (&errno, fd, (const char * )path, (char * )buf, bufsize),
-    /* _localcheck */ {if (!(cheri_getperm(path) & CHERI_PERM_GLOBAL)) {errno = EPROT; return ((int)-1);} if (!(cheri_getperm(buf) & CHERI_PERM_GLOBAL)) {errno = EPROT; return ((int)-1);} }
+    /* _localcheck */ {if (!(cheri_getperm(path) & CHERI_PERM_GLOBAL)) {errno = EPROT; return ((ssize_t)-1);} if (!(cheri_getperm(buf) & CHERI_PERM_GLOBAL)) {errno = EPROT; return ((ssize_t)-1);} }
 )
 
 SYS_STUB(501, int, renameat,
