+++ sys/compat/cheriabi/cheriabi_systrace_args.c
@@ -10275,7 +10275,7 @@ systrace_return_setargdesc(int sysnum, int ndx, char *desc, size_t descsz)
 	/* cheriabi_readlinkat */
 	case 500:
 		if (ndx == 0 || ndx == 1)
-			p = "int";
+			p = "ssize_t";
 		break;
 	/* cheriabi_renameat */
 	case 501:
